requirements.txt
psycopg2-binary
fastapi
pywebpush
uvicorn[standard]
langdetect
beautifulsoup4
starlette
databento
pydantic
py_vollib
lxml
scipy
gunicorn